Created CityService Interface and fixed Impl.
