+++ src/main/java/spotify/project/services/UserServiceImpl.java
@@ -26,7 +26,7 @@ import java.util.stream.Collectors;
 @Slf4j
 public class UserServiceImpl implements UserService {
 
-    private final CityService cityService;
+    private final CityServiceImpl cityService;
     private final UserRepository userRepository;
     private final RoleRepository roleRepository;
     private final PasswordEncoder passwordEncoder;
